add Popup window class
This adds popup windows that can be used for a variety of purposes, including:

    notifications (#1521)
    tooltips (#1441 )
    a keyboard overview window (#1329)

Once this is merged I can use it to add my notification server code into
qtile, which should make it clear if Popup needs updating e.g. to move some
of the setup logic into a ._configure method. I can then add a small bit in
the docstring to show how to use it for any future use.

Closes #1677