Use additional async functionality
Move more of the setup functionality into async. I think it'd be great if
nothing needed to be passed a loop in the constructor and all of the async
initialization could be called in async tasks, which would then be able to
automatically get the running loop. Some additional refactoring could help
with this. This may set us on a path that the Qtile object can be easily
re-made from within the session manager if we can pull out enough of the
direct loop handling from that.

Closes #1689